requirements.txt
futures<3.0.0
asyncpg==0.27.0
discord.py==2.1.0
ffmpeg
gTTS==2.2.3
Pillow
PyDictionary==2.0.1
random-word==1.0.11
psycopg2-binary==2.9.5
python-dotenv
PyNaCl
pyyaml
spotipy==2.17.1
SQLAlchemy==1.4.12